refactor: move `_mech_contract_interact` to base
